Initial implementation
This is a really rough implementation that has the basic functionality.  A navigation panel on the left hand side that pulls in the entire tree of metadata and then displays it so it can be navigated and clicking on each item uses HTMX to load in the value of that item into the right hand panel.

Overall this is still very rough, the goal is to have this be more of an informative tool for how the GCP Metadata store works so the actual content needing to be displayed on the right hand side needs to include much more information including formatting.

The biggest things missing are the ability to request an identity and access token