For each key showing the URL and the content returned for each key.
This is the point of this application.
The way that Google is handling these responses is not entirely intuitive.  There is a bunch of inconsistences here with respect to the cabob case as well as the response formats if it is text or JSON style therefore returning the payloads helps demonstrate how this works.