Added getting access token and identity token
This is the basic functionality, there are special cases for both the access token and identity token as they do not show up in the nominal JSON tree.

Therefore these are injected and also have special cases for rendering the pages.

Access token just displays the access token.

Identity token is more complex because the mandatory "audience" parameter must be specified.